Refactor config generation
First version had a separate code for each part to configure
 - one function for itsnet
 - one function for mw-server:
    - specific code to handle denservice.xml
    - specific code to handle positionprovider.xml
    - specific code to handle choirconf.xml

Replace with a more generic implementation. To achieve this,
slightly modify the JSON schema. Specifically, configuration variables
for each component (gn, DENM, position, CAM, ...) are grouped in a
separate object in the JSON message. This way, the only file that needs
special handling is choirconf.xml