Introduce gn_addr parameter to configure GN_ADDR of a station.
The gn_addr field in the JSON message must be in the ASCII format
accepted by itsnet. i.e. 4 16bit colon separated fields.

Sample valid GN_ADDR : 0000:4636:2ea8:87ee

This is copied in the itsnet.conf file. Moreover, the MAC address
of the interface is changed to match it.

If a gn_addr parameter is not provided, we keep a default value of
0:0:0:1. This in turn means that the GN_ADDR will be generated from
the MAC address of the eth0 interface to keep things coherent.