Add sample nginx conf for reverse proxy.
nginx proxies two kinds of connections :
 - to the control process (itseml), via the /ctl endpoint
 - to each env, via an endpoint of the form
    /env/<envid>/pubsub/(pub|sub)/(msgid)/(appid)

The proxy redirects the ctl endpoint to localhost. For each
environment, the IP has to be computed dynamically: to achieve
this, the embedded lua interpreter is used. On Debian, this means
that the nginx-extras package has to be installed rather than
the nginx/nginx-full one.