Generate configuration for TrafficLightService
This commit enables the SpatService, and adds code to parse a JSON
section to configure the TrafficLightService.
The format is the following:

  "trafficlight": {
     "states": [[3, 6, 7], [3, 6, 7]],
     "durations": [[10, 5, 10], [3,3,3]]
  }

An array defines a sequence of states for a traffic light.
There is a corresponding array that defines the duration of each
state in the "durations" section.

Since we support multiple traffic lights, both "states" and "durations"
are arrays of arrays.