Enable MAP support.
This commit enables the MapService in the middleware.
The MAP message is specified in the received JSON. It is written
out in the env's specific configuration directory. A service is
started to start the send_mapjson script that will read it and
trigger an emission at a rate of 1 message per second.