Modify MQTT host in publish_mqtt script
publish_mqtt attempted to publish on localhost, mostly as a placeholder
value, as the broker was supposed to be on some other device.

This has changed, and the broker is located on the machine hosting itseml,
albeit on the main routing table. As such, since the script is run from
within a netns, the broker's IP corresponds to the gateway of that netns.