Handle blinking lights in publish_mqtt.
The previous message format let us simply specify which light should
blink. With the new format, each byte is assumed to be in a binary
state. Thus, the blinking is handled by the script: request a light to
be switched on, then switch it off.

While here, operate on new event transitions rather than sending a
message for each SPAT processed. This helps reduce the number of sent
messages and handle an issue that stems from the blinking light handling.

Specifically, when we handle a blinking light, subsequent messages
are buffered, which means a burst is going to happen on the next transition.