Y-Ghost : add files to implement a light middleware instance
These files are for a light middleware instance which is launched each time where we want to deploy a new C-ITS virtual instance.

in mw-server/, we find necessary find to the proper functionnement of the light middleware
service_file/license-mw-server.service is a systemd file to launch the light middleware as a service.

However, a .p12 certificate under the name yogoko-middleware.p12 will have to be added.

Signed-off-by: Léo <leo.lestanguennec@yogoko.fr>