Y-Ghost : update itseml config files to integrate download-static license support
In itseml/itseml.py,a function to verify the integrity of the yogoko license or certificate.
If the port 49154 is open, this means the light middleware instance has passed his license check.
For each C-its virtual instance deploy, the light middlware instance restart to check the integrity of the license.

In itseml/templates/mw/choirconf.xml, the license section is added to enabled license check in order to evitate false-positive.

In setup.py, the files of previous commit are add in the itseml installation.

Signed-off-by: Léo <leo.lestanguennec@yogoko.fr>