Allow the user to easily spawn actors in a CARLA simulation.
The manual_control.py script is an example script from the CARLA simulator which allows the user to spawn a vehicle in an existing simulation and controll it manually with the keyboard.
The spawn_npc.py script is a custom script that allows the user to spawn NPC vehicles and pedestrians in an existing simulation.

Signed-off-by: Lola <lola.umenhover@yogoko.fr>