Fix wrong license file path
The way the path to the template directory was referred is wrong, as the
actual path is /usr/lib/ and not /usr/local/lib/ when building the debian
package. Moreover, since the path to the template directory is already defined
in the python script as a variable, defining it statically is also wrong.

More generally, /usr/lib/ and /usr/local/lib/ folders shall be used to store
**static** data. Variable data shall be placed in /var/. As a consequence,
this directory will need to be changed again in a later commit.

Signed-off-by: Emmanuel Thierry <emmanuel.thierry@yogoko.fr>